Selecting Name Change and CSS Updates
Instead of using native Select/Option, I  prefer when a popover
dropdown select menu is used.

So user click on Player name box and either Ul#homePlayers or
Ul#awayPlayers is shown. These would be hidden until one is click.

I could try and implement it fully, but thought it best to check to see
whether you are happy to go ahead with this, and it might be quicker if
you did it.